+++ src/app/auth/login/form.tsx
@@ -1,17 +1,26 @@
 'use client';
 
+import { useState } from 'react';
 import Link from 'next/link';
 import { Controller } from 'react-hook-form';
+import { Eye, EyeOff } from 'lucide-react';
 
 import { Button } from '@/components/ui/button';
 import { Field, FieldGroup, FieldLabel, FieldError } from '@/components/ui/field';
 import { Input } from '@/components/ui/input';
+import { InputGroup, InputGroupAddon, InputGroupButton, InputGroupInput } from '@/components/ui/input-group';
 import { Spinner } from '@/components/ui/spinner';
 
 import useLoginForm from './use-form';
 
 export function LoginForm() {
   const { form, onSubmit, isPending } = useLoginForm();
+
+  const [passwordShow, setPasswordShow] = useState(false);
+
+  const handlePasswordShow = () => {
+    setPasswordShow((prevState) => !prevState);
+  };
 
   return (
     <form onSubmit={form.handleSubmit(onSubmit)} className="'flex gap-6' flex-col">
@@ -47,14 +56,27 @@ export function LoginForm() {
                   Forgot your password?
                 </Link>
               </div>
-              <Input
-                {...field}
-                id="password"
-                type="password"
-                aria-invalid={fieldState.invalid}
-                placeholder="••••••••"
-                autoComplete="off"
-              />
+              <InputGroup>
+                <InputGroupInput
+                  {...field}
+                  id="password"
+                  type={!passwordShow ? 'password' : 'text'}
+                  aria-invalid={fieldState.invalid}
+                  placeholder="••••••••"
+                  autoComplete="off"
+                />
+                <InputGroupAddon align="inline-end">
+                  <InputGroupButton
+                    aria-label={!passwordShow ? 'Show password' : 'Hide password'}
+                    title={!passwordShow ? 'Show password' : 'Hide password'}
+                    size="icon-sm"
+                    className="cursor-pointer"
+                    onClick={handlePasswordShow}
+                  >
+                    {!passwordShow ? <Eye /> : <EyeOff />}
+                  </InputGroupButton>
+                </InputGroupAddon>
+              </InputGroup>
               {fieldState.invalid && <FieldError errors={[fieldState.error]} />}
             </Field>
           )}