+++ CHANGELOG.md
@@ -53,6 +53,10 @@ All notable changes to this project will be documented in this file.
 - posts authors schemas
 - posts authors store
 
+##### Auth
+
+- login form password show/hide button
+
 #### Components
 
 - approve and reject buttons
@@ -74,7 +78,7 @@ All notable changes to this project will be documented in this file.
 #### Components
 
 - list search schema to allow strings
-- list search form input and button to a shadcn input group
+- list search form input and button to an input group
 
 ## [0.3.0] - 2026-04-05